refactor: group impl classes
