+++ src/main/java/com/safetynet/alerts/dao/impl/MedicalRecordDaoImpl.java
@@ -1,5 +1,6 @@
-package com.safetynet.alerts.dao;
+package com.safetynet.alerts.dao.impl;
 
+import com.safetynet.alerts.dao.MedicalRecordDao;
 import com.safetynet.alerts.dto.alerts.DataDto;
 import com.safetynet.alerts.model.MedicalRecord;
 import org.springframework.beans.factory.annotation.Autowired;
