+++ src/test/java/com/safetynet/alerts/service/AlertsServiceTest.java
@@ -7,6 +7,7 @@ import com.safetynet.alerts.dto.alerts.*;
 import com.safetynet.alerts.model.FireStation;
 import com.safetynet.alerts.model.MedicalRecord;
 import com.safetynet.alerts.model.Person;
+import com.safetynet.alerts.service.impl.AlertsServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
