+++ src/test/java/com/safetynet/alerts/service/PersonServiceTest.java
@@ -5,6 +5,7 @@ import com.safetynet.alerts.dto.person.DeleteBody;
 import com.safetynet.alerts.dto.person.PostBody;
 import com.safetynet.alerts.dto.person.PutBody;
 import com.safetynet.alerts.model.Person;
+import com.safetynet.alerts.service.impl.PersonServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
